# api.yaml
openapi: "3.1.0"
...
servers:
  - url: http://localhost:3000
    description: Local Server
  - url: https://patiofy.smartaihr.com
    description: Production Server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
info: 
  title: Patiofy website Api
  description: API for patiofy website
  contact:
    name: support team
    url: http://Patiofy.com/contactUs
    email: ajaymaroju@techpixe.com
  termsOfServices: http://Patiofy.com/terms
  licence:
    name: MIT Licence
    url: http://opensource.org/licenses/MIT
paths:
...
  /patiofy/auth/user/google/{id}:
    put:
      tags:
        - Google_OAuth2.0
      summary: Set a new password for a user
      description: Update the password for a user based on their ID.
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the user whose password is being updated.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                newpassword:
                  type: string
                  example: Ajay@110125
      responses:
        '200':
          description: Password updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: password updated successfully
        '400':
          description: Missing or invalid password
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: password is required
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Internal Server Error
                  error:
                    type: object
  /patiofy/auth/user/signup:
    post:
      summary: User Signup
      description: Creates a new user account. Sends a verification email after successful signup.
      tags:
        - users
      security:
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - firstname
                - lastname
                - email
                - password
              properties:
                firstname:
                  type: string
                  example: Ajay
                lastname:
                  type: string
                  example: Maroju
                email:
                  type: string
                  format: email
                  example: marojuajay111@gmail.com
                password:
                  type: string
                  format: password
                  example: Ajay@110125
                istermsandConditions:
                  type: boolean
                  example: true
      responses:
        '200':
          description: User created and verification email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: verification has been send to the email, please verify
        '401':
          description: Email already taken
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: email already taken, try another
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Internal Server Error
                  error:
                    type: string
                    example: Some error occurred
  /patiofy/auth/user/verify:
    get:
      summary: Verify user account
      description: |
        This endpoint verifies a user's account by decoding the verification key sent as a query parameter. 
        It checks if the verification time has expired and updates the user's status to "active" upon successful verification.
      tags:
        - users
      parameters:
        - name: verificationKey
          in: query
          required: true
          schema:
            type: string
          description: The base64 encoded verification key sent to the user's email.
      responses:
        "200":
          description: Account verified successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Account verified successfully"
        "408":
          description: Verification time expired.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Time expired, please register"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal Server Error"
  /patiofy/auth/user/resend:
    post:
      summary: Send account verification email
      tags:
        - users
      description: Sends an account verification link to the specified email if the user exists.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
      responses:
        '200':
          description: Verification link sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: link send to the email successfully
        '400':
          description: Missing email in request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: email is required
                  error:
                    type: string
                    example: Bad Request
        '404':
          description: User not found with given email
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: user not found
                  error:
                    type: string
                    example: Not Found
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Internal Server Error
                  error:
                    type: string
                    example: Some server error details
  /patiofy/auth/user/resendset:
    post:
      tags:
        - users
      summary: Resend password set link
      description: Sends a password setup email to the user if their password is not already set.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
      responses:
        '200':
          description: Link sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  statuscode:
                    type: integer
                    example: 4
                  message:
                    type: string
                    example: link send to the email successfully
        '400':
          description: Email not provided
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  statuscode:
                    type: integer
                    example: 1
                  message:
                    type: string
                    example: Email is required
                  error:
                    type: string
                    example: Bad Request
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  statuscode:
                    type: integer
                    example: 2
                  message:
                    type: string
                    example: user not found
                  error:
                    type: string
                    example: Not Found
        '402':
          description: Password already set
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  statuscode:
                    type: integer
                    example: 3
                  message:
                    type: string
                    example: Password has already been set, please login
                  error:
                    type: string
                    example: Invalid Action
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  statuscode:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: Internal Server Error
                  error:
                    type: string
                    example: Error message details
  /patiofy/auth/user/signin:
    post:
      tags:
        - users
      summary: Sign in a user
      description: Authenticates a user with email and password, and returns a JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: marojuajay111@gmail.com
                password:
                  type: string
                  example: Ajay@110125
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: login successfully
                  JWTtoken:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '401':
          description: Authentication failed due to missing/invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: incorrect Email or password does not match
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Internal Server Error
                  error:
                    type: string
  /patiofy/auth/user/me/{id}:
    get:
      summary: Get user by ID
      description: |
        This endpoint retrieves a user's details by their unique ID.
      tags:
        - users
      security:
        - BearerAuth: [] 
      responses:
        "200":
          description: User details retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    type: object
                    description: The user object containing details.
                    properties:
                      _id:
                        type: string
                        example: "65a2d5e2f1c3a12b34f6a789"
                      name:
                        type: string
                        example: "John Doe"
                      email:
                        type: string
                        example: "johndoe@example.com"
        "404":
          description: User not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "User not found"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Internal Server Error"
  /patiofy/auth/user/update:
    put:
      summary: Update User Profile
      description: Updates the logged-in user's first name and last name.
      tags:
        - users
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstname:
                  type: string
                  example: John
                lastname:
                  type: string
                  example: Doe
                phone:
                  type: number
                  example: 9000000000
                Address:
                  type: string
                  example: warangal, Telangana, India
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Profile updated successfully
                  updatedUser:
                    type: object
                    properties:
                      _id:
                        type: string
                        example: 609c0f58f3a9a2001510bdf1
                      firstname:
                        type: string
                        example: John
                      lastname:
                        type: string
                        example: Doe
                      email:
                        type: string
                        example: john.doe@example.com
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Internal Server Error
                  error:
                    type: string
                    example: Some error occurred
  /patiofy/auth/user/profilepic:
    post:
      summary: Upload or update user profile picture
      tags:
        - users
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              image:
                - profilePic
              properties:
                image:
                  type: string
                  format: binary
                  description: User profile image file
      responses:
        '200':
          description: Profile picture uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  statuscode:
                    type: integer
                    example: 3
                  message:
                    type: string
                    example: User profile updated successfully
        '401':
          description: No file provided
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  statuscode:
                    type: integer
                    example: 1
                  message:
                    type: string
                    example: profile image is required
                  error:
                    type: string
                    example: Bad Request
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  statuscode:
                    type: integer
                    example: 2
                  message:
                    type: string
                    example: User not Found
                  error:
                    type: string
                    example: Not Found
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  statuscode:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: Internal Server Error
                  error:
                    type: string
                    example: error.message
  /patiofy/auth/user/password/forget:
     post:
      summary: Request password reset
      description: Sends a reset password link to the provided email address.
      tags:
        - users
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: marojuajay111@gmail.com
      responses:
        '200':
          description: Password reset link sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: reset password link sent to the email
        '401':
          description: Incorrect email or validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: incorrect email
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Internal Server Error
                  error:
                    type: string
                    example: Detailed error message
  /patiofy/auth/user/password/setNew:
    put:
      summary: Set a new password for a user
      tags:
        - Google_OAuth2.0
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - newpassword
              properties:
                email:
                  type: string
                  format: email
                  example: marojuajay111@gmail.com
                newpassword:
                  type: string
                  format: password
                  example: Ajay@110125

      responses:
        '200':
          description: Password updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: new password updated , please login
        '400':
          description: Missing email or new password
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: All fields are required
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Internal server Error
                  error:
                    type: string
                    example: Error message details
  /patiofy/auth/user/password/setforget:
    post:
      tags:
        - users
      summary: Set or update a new password
      description: Allows a user to set a new password using their email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - newpassword
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                newpassword:
                  type: string
                  format: password
                  example: MyNewSecurePassword123!
      responses:
        '200':
          description: Password updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  statuscode:
                    type: integer
                    example: 4
                  message:
                    type: string
                    example: password updated successfully, please login
        '400':
          description: Required fields missing
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  statuscode:
                    type: integer
                    example: 1
                  message:
                    type: string
                    example: all fields are required
                  error:
                    type: string
                    example: Bad Request
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  statuscode:
                    type: integer
                    example: 2
                  message:
                    type: string
                    example: User not found
                  error:
                    type: string
                    example: Not Found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  statuscode:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: Internal Server Error
                  error:
                    type: string
                    example: Detailed error message
...
  /patiofy/auth/user/delete:
    delete:
      summary: Delete user account
      description: |
        This endpoint allows a user to delete their account. 
        The user must provide their password for verification. 
        If the password is correct, the account is deleted, and the authentication token is cleared.
      tags:
        - users
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  description: The user's password for verification.
                  example: Ajay@110125
      responses:
        "200":
          description: User account deleted successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "User account deleted successfully"
        "401":
          description: Password is required or incorrect.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "password is required or Incorrect password"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal Server Error"
...